Pre-filter Sentry data before it is sent over the wire
While we'll also be scrubbing PII in SaaS Sentry, it makes sense to never let
PII leave our servers. Bedrock has very, very little behaviour that handles
PII, but we're pre-scrubbing to be safer.

Note that the two fields scrubbed are based on the current self-hosted Sentry
configuration for Bedrock.

Includes a similar smoke test to Basket, to be sure that sentry_processor is
scrubbing what we expect, but note that it's not exhaustively tested.