Update Pocket locale/L10N config based on latest info about translations from Vendor
We'll be getting/using "es-ES" not just "es" in our interactions with the vendor
which means we will need to support some kind of redirect/rewrite at the CDN
level. Details of this TBC, but with the current config in this changeset,
Bedrock WILL handle the locale differences (redirecting from /es/foo to
/es-ES/foo, for example) but will do so with a HTTP 302 Temporary Redirect
which has an SEO impact. Ideally any redirect because the locale code has been
tweaked must be a 301.

Also, it's not just es-ES and es-la/LA that need attention. Comparing with
https://github.com/Pocket/dotcom-gateway/blob/main/lambda/src/config/locales.ts
/zh/ and /pt/ will also need redirects