Stop the mandala animation on the whatsnewpage
This keeps wasting CPU cycles whenever my Firefox Developer Edition updates.

Part of the reason is that I have two Firefox windows. If I open FIrefox but I'm focused on the wrong window, then I don't see this page. I just hear my laptop's CPU spinning, and wonder why.

There were 6 upvotes when I raised [this issue on Mozilla Connect](https://connect.mozilla.org/t5/ideas/do-not-put-a-continuous-animation-on-the-whatsnew-page-or-stop/idi-p/7297), so I guess I'm not the only person bothered by this.

A better solution might be to disable the animation when the page is not focused. We could listen for the `blur` event, but I'm not sure if that will fire if Firefox starts up with multiple windows.