Specify a downpinned version of the CircleCI image to avoid CI failures
As of 5 Oct, our CI pipeline started seeing test errors unrelated to
code changes in _any_ particular PR. These failures were not replicable
locally (bare metal MacOS, or via Docker) and were not replicable via a
SSH session to CircleCI, despite using the same images that had just failed.

My theory is that there is something about how Bedrock interacts with the
3.9.14 version of Python in the Circle container that causes
the very odd failures (eg Mocks not returning the expected values). We will
continue to explore this, but this changeset downpins us to using an earlier
cut of Circle's Python 3.9 image, which we know works. The vulnerabilities
fixed in 3.9.14 seem to pose a low risk, given that we are not building
production images here, and the actual Bedrock Docker images are using 3.9.14
-- it's just the "outer" image which is being downpinned, and ideally only
temporarily.