Pocket: update /my-list and "My List" to /saves and "Saves"
This change is designed to match in-app changes for main Pocket.

The wording and URL change is overseen by a switch:

SWITCH_POCKET_RENAME_MY_LIST_TO_SAVES

A follow-up task is required to clean this up after go-live.

We need to do the work slightly early so that there's time to get the
translations all in place and tested.