Only bleach the HTML in the raw markdown content, not the rendered HTML
b29693789ecab21938fc24cc49f56dc8874bd2a2 was an attempt to improve sanitisation
of HTML from markdown, but doing so inadvertently over-escaped some of the
HTML tags that were produced by rendering markdown (eg img) because they
were not in the allowlist.

Revisiting this, I think it's OK to revert to only sanitising the raw
markdown's HTML elements, because any rendered HTML from the markdown is
only going to be adding safe HTML tags, based on the spec.