Update Pocket GTM Container (fixes #12624)
We have moved Pocket's GA4 property under the Mozilla Account.
We use a lookup table in the GTM container to set measurement ID
based on whether the hostname is production or not.
If production, the measurement ID connects to getpocket.com - GA4
property.
If not production, the measurement ID connects to getpocket.com -
GA4 Testing property.