Article theme style fixes
Content background color was used in fixed header styles
(which do not apply for reduced motion prefs). This creates a
visual bug for reduced motion users. It is better placed with
theme styles because it is always needed to override the theme
background color on content.

This also fixes a bug where the color cuts off unexpectedly if
the viewport height is shorter than the header.