Port CDN tests to Github Actions (#13110)
This changeset reimplements the CDN tests that we had on GitLab as a GH Action.

Note that the GL tests were run with the "cdn and not cdnprod" Pytest mark,
which I've taken to mean the single 'cdnprod' test in Bedrock was not run
(it is an SSL check). As such, I run the CDN tests with "cdn and not cdnprod"
once and then also run them with just the "cdnprod" mark so ensure that's run
too.