Refactor redirect for Pocket Updates Signup pilot
Unfortunately, due to a slightly convoluted circular dependency, we have to
exclude the bedrock.redirects app from Pocket Mode, meaning we have to do
redirects via the 'manual' way, rather than using the neat helper. This
changeset does that work so we can ship this work

Background:

Pocket mode has to exclude bedrock.redirects because of a redirect in
bedrock.mozorg that clashes with a route avaialble in Pocket Mode (/jobs/).
We can't drop just bedrock.mozorg from INSTALLED_APPS in Pocket Mode
because there is a helper (resp_img) which is needed in Pocket Mode. We _can_
refactor our way out of this, but it'll need a bit of planning and is better
as a standalone fix.