Fix VPN redirects (Fixes #13460) (#13712)
* Fix VPN redirects (Fixes #13460)

Re-adds pages at vpn/more/* which were removed in #12938

Adds a view in products/views.py that takes three arguments (template, resource-center slug, and ftl_files) - this way, when navigating to a vpn/more/* url - this view checks the associated page in the VPN resource center and checks if the active locale is available. If it is available, the user is redirected to the resource center, if not, the user is sent to the correct local with the vpn/more/* url.