Fix issue with unclosed img tag in Nightly Feed
This changeset deals with a problem using Markdown to render an XML feed for
Nightly. Specifically, we were seeing an <img> tag that was not being closed,
which then breaks XML parsing, meaning RSS readers cannot show the content.

The changeset adds a new postprocessor that looks for an unclosed <img> tag
and closes it.

Rather than change our overall Markdown rendering, the fix has beebn isolated
to _just_ the Nightly feed.

Resolves #13830
Resolves https://bugzilla.mozilla.org/show_bug.cgi?id=1863200