Remove links to Mozillians.org from leadership pages
These links seem pretty old and are of low value because Mozillians.org now
redirects to people.m.o which in turn needs a user to be authenticated in order
to see profiles.

Mark's link redirected to a valid people.m.o page, but behind auth
Angela and Ashley's profiles had moved to a different slug on people.m.o.

Also, the link for Mitchell's profile was wrong and was pointing to one with
a similar slug but not her actual people.m.o profile

Resolves #14059