First part of adding in Wagtail
Bedrock uses Jinja2 templates via the django-jinja package. Wagtail does support
jinja2 templates, but with a different template engine that doesn't seem to
support the extensions we want (at least not without some changes). So, we've
added a custom template backend that it scoped to Wagtail templates - and if it
finds those, it uses the regular Django Template Language (DTL) parser for them

We also only enable the Wagtail admin if a particular setting is truthy, which
means we can have a single editing instance/host while keeping the web pods
read-only.

I'm trying to keep the need for Sessions out of the main flow as much as
possible, so while the SessionMiddleware is now added to settings.MIDDLEWARE
(so that the Wagtail admin can use proper auth), we don't end up with a
needless session cookie for regular, read-only unauthenticated views.

Also auto-formats a lot of settings while in the neighbourhood via ruff