Add a management command that will send an email to check deliverability
If we set EMAIL_DELIVERABILITY_TEST_ADDRESS to be the address of a
dedicated snitch in Dead Man's Snitch and then run it on a k8s cron
job, in production only, we will be able to be alerted if email
deliverability is affected.

There will be a corresponding infra PR that uses this management command.

It was kept out of cron.py for simplicity in controlling exactly
where it is triggered.

I used Copilot to generate the base of the tests, but then tuned them by hand
(to make them appropriate, not just to make them work)