Add helper to get only the hostname out of MEDIA_URL
settings.MEDIA_URL (passed in here as media_url) is a custom route on our CDN
that points to a cloud bucket, which we use for uploaded media from the CMS.

Specifically, due to infra constraints, it has to point to a _sub-path_ in the bucket,
not the top/root of it. It also needs to be distinct from the CDN route that points to our
collected static assets (which are at https://<CDN_HOSTNAME>/media/ - that is STATIC_URL)
With all this in mind, MEDIA_URL, when set, points to https://<CDN_HOSTNAME>/media/cms/
When django-storages computes the URL for a object in that CMS bucket, it wants
just the hostname of the bucket, not the full CDN/proxy path to the subdir in the bucket,
because it opinionatedly concatenates it with GS_LOCATION (which is needed to ensure
the files are uploaded to the sub-path mentioned above).

TLDR: We just need the root of the CDN, from MEDIA_URL.