Update healthz-cron page re postgres backend
Bedrock will continue to operate with sqlite for local dev, demos and
bedrock-test, but otherwise will use postgres as its DB backend.

As such, some of the info on the healthz-cron page will be misleading if
postgres is in use, because it's focused on sqlite.

This changeset hides the sqlite-specific data if sqlite is not in use.
It also adds a new timestamp that shows the most recent timestamp of
all data sources that have been pulled in, as a rough way for us to
quickly see that the run-db-update job is still working. (There's also
a Dead Man's Snitch watching the relevant job, but this adds a
convenient quick look)