Refactor sqlite DB export script
Usage: DATABASE_URL="postgres://user:pass@host:5432/bedrock" ./bin/export-db-to-sqlite.sh /path/to/output.db

Previously we were dumping most of the DB by specifying what to exclude. Now
we've flipped it explicitly including the data we want, which will make
the risk of accidental inclusion of irrelevant or sensitive data less likely
if a new Django app is added to the project, or a model grows to add new fields.

We do still need to temporarily export some tables from postgres that we don't
want to retain in the target sqlite DB (eg the auth_user table) - this is only
so that we can load in the data into the target DB without integrity issues.
We then null out references to such tables before clearing them.

Still TODO:
* make it easier to expand the tables being de-linked and cleared, to avoid
bugs in the future
* improve the error detection behaviour so we can fail hard if something does
not work
* decide how to fit this into the bigger picture of exporting a DB that will
get uploaded to the cloud bucket