Bugfix: ensure correct list of available translations for a CMS-backed page (#15062)
By default, when a page is added, wagtail-localize will add an Alias per locale
for each available locale, unless an actual translation of that page exists.
This meant that, prior to this changeset, the languge selector in the footer
was including languages which only had an Alias of a page available - and
an Alias just links to the en-US content, but with a path of, say /fr/...
which is misleading.

There'll be separate work to stop the auto-generation of Alias pages by default
but as an extra safeguard, this changeset deliberately ignores Alias pages
when computing the list of actual locales available for a CMS-backed page.