Ensure CMS StructuralPage pages don't appear in the sitemap unless they are also parents of real pages
In the CMS we use a StructuralPage page type to create branches/folders in the
page tree.

These StructuralPages do not have content but allow us to structure URL paths
as we need, including matching some routes that already exist in the page tree
as static/Django-only paths

However, we've spotted a bug: prior to this changeset, if there's a static path
for foo/ and a StructuralPage for foo/, the static paths for foo/ are dropped
from the sitemap data in favour of those from the CMS _even if the CMS has fewer
locales set up than the static route_. This can be seen here where /about/
was in the CMS as StructuralPage only in en-US, but in lots of locales as a
static page.

https://github.com/mozmeao/www-sitemap-generator/commit/bdd1cc0fcb0d62ada6866479aea88018041f9ec4

This changeset fixes things by ensuring we ignore StructuralPages when building
page URLs from the CMS, unless they are parents of non-StructuralPages.

The tests show an example of this.