Fix fatal error when exporting a source DB
...where the auth_user_groups and/or auth_user_user_permissions tables have data

These two tables have a non-NULL user_id column, which means that previously
the script was failing because it could not nullify rows in those tables.

The data for those rows was being pulled over via FK constraints because
we exported the auth.User model from the source DB.

The fix was to ensure that we were actually purging those two tables
as well as the auth_user table (which we were doing already).

Also, the ordering of operations needed to be addressed: clearing the tables
before running the nullifying-relations code, rather than after, to avoid
integrity issues because of that un-nullable auth_user_groups.user_id.

I've tested this locally and putting the nulling after the table deletion
seems to be fine: nullable columns are nulled and no data appears to
cascade-delete when auth_user, auth_user_groups or auth_user_user_permission
is dropped. If cascading deletion was going to occur, we would already
have seen it with the deletion of auth_user, regardless of when it's called
in the script.