Ensure pre-commit uses LTS version of node
Since Node 23 came out, fresh installs or updates to our pre-commit hooks
will fail for eslint. This is because pre-commit will pick the very new
version of Node, because we haven't told it not to.

This change tells pre-commit to use the LTS version installed on the system,
which will match what we use for main Bedrock