Remove strikethrough as an option for rich text
Struck-through text does not end up as editable at our l10n vendor's side of
things, which means localization breaks for any page containing it.

We don't use it anyway, so better to avoid it being sent over altogether.

The migration is needed because it describes the state of the page model
that already has a rich-text field in it, and we just changed that state
by removing a formatting option.

Resolves #15498