Add helpers to support a 'hybrid cache' option that uses locmem + DB cache
In order to balance the need for a distributed cache with the speed of a
local-memory cache, we've come up with a couple of helper functions that wrap
the following behaviour:

Getting values:
* If it's in the local-memory cache, return that immediately.
* If it's not, fall back to the DB cache, and if the key exists there,
return that, cacheing it in local memory again on the way through
* If the local memory cache and DB cache both miss, just return the
default value for the helper function

Setting values:
* Set the value in the local memory cache and DB cache at (almost) the same time
* If the DB cache is not reachable (eg the DB is a read-only replica), log
this loudly, as it's a sign the helper has not been used appropriately, but
still set the local-memory version for now, to prevent total failure.

IMPORTANT: before this can be used in production, we need to create the
cache table in the database with ./manage.py createcachetable AFTER this
code has been deployed. This sounds a bit chicken-and-egg but we hopefully can
do it via direct DB connection in the worst case.