Support failing over to a different, preferred locale for CMS pages upon 404
This changeset addresses the problem with a request to path in a locale
that is not in the CMS does not fall back to a locale variant for that
path that _is_ in the CMS.

e.g. if /en-US/about/leadership/ exists in the CMS but /fr/about/leadership
does not, typical Bedrock behaviour (for non-CMS pages) would be to
redirect to the /en-US/ version as a fallback.

This changeset does more than just punt to en-US, though - it aims to
redirect to the best-fit locale available in the CMS based on the
user's Accept-Language header, and then will fall back to trying en-US

Fixes #15932