(fix) Ensure we always select a specific release for ESR relnotes
This is because we now have more than one ESR available at any one time.

We use the product_details lib as the source of truth for which version
we want for the "esr" channel, because we can't depend on the ESR versions
being saved/created in a reliable order.

Manually testable by clicking the 'Release notes' link on http://localhost:8000/en-US/firefox/all/desktop-esr/

Resolves #16289