Remove the old IT-owned Mozilla Root certificates. (#16595)
This is effectively a revert of 6edb8a88dfde7ba60eebc625df57b97633cae35d / bug 1149019

The two certs here are expired: the SHA1 one since 2016, the SHA2 since 2025.  I became owner of them through attrition.  We stopped issuing from them in 2018, and held on to the relics until expiration.  At this point, the are all expired, the private keys have been irrevocably destroyed, and we have no intention of making a new one, ever.