Update node to v24 and set npm version in node engines
Dependabot was using a buggy version of npm that messed up our lock
files. The use of engines field in package.json should guide dependabot
to use the correct npm version.

Developers can use n or nvm to keep their node version in sync with the
engines field, but this is not strictly enforced.

We are moving to node v24 across docker and workflows as a more general
maintenance update.